requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
openai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
